requirements.txt
streamlit==1.45.0
boto3==1.38.8
openai==1.77.0
langchain==0.3.25
qdrant-client==1.14.2
python-dotenv==1.1.0
langchain-openai==0.3.16
langchain-core==0.3.58
PyJWT==2.10.1
plotly==6.0.1
cryptography==44.0.3


